BagWeight & NoOfBag in billing and resettlement & company selection problem for billing & resettlement done

+++ SA.HM.Data/Restaurant/KotBillDetailDA.cs
@@ -670,8 +670,8 @@ namespace HotelManagement.Data.Restaurant
                                 entityBO.CategoryId = Convert.ToInt32(reader["CategoryId"]);
                                 entityBO.UnitHead = reader["UnitHead"].ToString();
                                 entityBO.Remarks = reader["Remarks"].ToString();
-                                entityBO.BagWeight = Convert.ToInt32(reader["BagWeight"]);
-                                entityBO.NoOfBag = Convert.ToInt32(reader["NoOfBag"]);
+                                entityBO.BagWeight = Convert.ToDecimal(reader["BagWeight"]);
+                                entityBO.NoOfBag = Convert.ToDecimal(reader["NoOfBag"]);
 
                                 entityBO.ItemType = reader["ItemType"].ToString();
                                 entityBO.ItemId = Convert.ToInt32(reader["ItemId"]);
@@ -778,8 +778,8 @@ namespace HotelManagement.Data.Restaurant
                                     entityBO.VatAmount = Convert.ToDecimal(reader["VatAmount"]);
                                 }
                                 entityBO.ItemCost = Convert.ToDecimal(reader["ItemCost"]);
-                                entityBO.BagWeight = Convert.ToInt32(reader["BagWeight"]);
-                                entityBO.NoOfBag = Convert.ToInt32(reader["NoOfBag"]);
+                                entityBO.BagWeight = Convert.ToDecimal(reader["BagWeight"]);
+                                entityBO.NoOfBag = Convert.ToDecimal(reader["NoOfBag"]);
 
                                 entityBOList.Add(entityBO);
                             }
